# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Agriculture API Docs",
        "description": "API Docs for Agriculture Reports",
        "contact": {
            "email": "sajeeshe@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Generate Tokens",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Generated Token"
                    }
                }
            }
        },
        "/api/fields": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "List All Fields",
                "responses": {
                    "200": {
                        "description": "Return Field list"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/field/add": {
            "post": {
                "tags": [
                    "Fields"
                ],
                "summary": "Add new Field",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Field name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "area",
                        "in": "query",
                        "description": "Area",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "crops_type",
                        "in": "query",
                        "description": "Crop Types",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Created Field"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/field/{slug}": {
            "get": {
                "tags": [
                    "Fields"
                ],
                "summary": "Display individual Field detail",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Field Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return selected Field details"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/field/update/{slug}": {
            "put": {
                "tags": [
                    "Fields"
                ],
                "summary": "Add new Field",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Field name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "area",
                        "in": "query",
                        "description": "Area",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "crops_type",
                        "in": "query",
                        "description": "Crop Types",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Field Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Updated Field"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/field/delete/{slug}": {
            "delete": {
                "tags": [
                    "Fields"
                ],
                "summary": "Delete a Field",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Field Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Deleted Tractor details"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}